refactor: rename files for clearer representation
- config.go → config_loader.go
- server.go → config.go
